+++ packages/vue-stack/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @fastkit/vue-stack
 
+## 0.15.49
+
+### Patch Changes
+
+- Added the `resolveHandler` option for performing custom validation and asynchronous processing in the stack's resolve process, along with the `guardInProgress` getter to determine if asynchronous processing is in progress.
+
+  Additionally, fixed an issue where the stack was hidden even on failed navigation when `closeOnNavigation` was set to `true`.
+
 ## 0.15.48
 
 ### Patch Changes